update of works until presentation of 23/06/2021

+++ src/main/java/io/grakn/pcrtests/migrate_csv_to_grakn/InputDepartements.java
@@ -22,7 +22,7 @@ public class InputDepartements extends InputSetOfData {
 		super(optionImport, filename);
 	}
 	
-	public InputDepartements(OptionImport optionImport, String filename, int myMaxGet, int minIndice, int maxIndice)
+	public InputDepartements(OptionImport optionImport, String filename, int myMaxGet, Long minIndice, Long maxIndice)
 	{	
 		super ( optionImport, filename,  myMaxGet,  minIndice,  maxIndice);
 	}
