+++ src/main/java/io/grakn/pcrtests/migrate_csv_to_grakn/InputDepartementsLinks.java
@@ -27,7 +27,7 @@ public class InputDepartementsLinks extends InputSetOfData {
 		super(optionImport, filename);
 	}
 	
-	public InputDepartementsLinks(OptionImport optionImport, String filename, int myMaxGet, int minIndice, int maxIndice)
+	public InputDepartementsLinks(OptionImport optionImport, String filename, int myMaxGet, Long minIndice, Long maxIndice)
 	{	
 		super ( optionImport, filename,  myMaxGet,  minIndice,  maxIndice);
 	}
